pyproject.toml
[project]
name = "code-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.63.0",
    "python-dotenv>=1.0.1",
    "tree-sitter>=0.24.0",
    "tree-sitter-python>=0.20.2",
    "tree-sitter-javascript>=0.20.1",
    "tree-sitter-typescript>=0.20.3",
    "tree-sitter-ruby>=0.19.0",
    "tree-sitter-go>=0.19.1",
    "tree-sitter-java>=0.20.2",
]
